4th nov(error)

+++ shoppingbackend/src/main/java/com/shoppingbackend/dto/Cart.java
@@ -34,7 +34,7 @@ public class Cart implements Serializable {
 	
 	
 	@Column (name = "grand_total")
-	private int grandTotal;
+	private double grandTotal;
 	@Column (name = "cart_lines")
 	private int cartLines;
 	
@@ -47,10 +47,11 @@ public class Cart implements Serializable {
 	public void setId(int id) {
 		this.id = id;
 	}
-	public int getGrandTotal() {
+	
+	public double getGrandTotal() {
 		return grandTotal;
 	}
-	public void setGrandTotal(int grandTotal) {
+	public void setGrandTotal(double grandTotal) {
 		this.grandTotal = grandTotal;
 	}
 	public int getCartLines() {
